refactor: model 내의 클래스 세분화 하여 패키지 위치 이동
model 하위 class10개 이상으로 추상화 정도에 따라 나누어 경로 이동시켜줌.
