+++ src/main/java/model/ladder/Height.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 public class Height {
     private static final int MINIMUM_HEIGHT_LIMIT = 1;
