+++ src/main/java/model/prizes/Prizes.java
@@ -1,4 +1,4 @@
-package model;
+package model.prizes;
 
 import java.util.Arrays;
 import java.util.List;
