+++ src/main/java/model/prizes/RewardBoard.java
@@ -1,8 +1,9 @@
-package model;
+package model.prizes;
 
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
+import model.people.Person;
 
 public class RewardBoard {
     private final Map<Person, Prize> board;
