+++ src/test/java/model/ladder/HeightTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
