+++ src/test/java/model/ladder/LadderTest.java
@@ -1,7 +1,8 @@
-package model;
+package model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import model.Direction;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
