+++ src/test/java/model/ladder/LineTest.java
@@ -1,8 +1,9 @@
-package model;
+package model.ladder;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import model.Direction;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
