+++ src/test/java/model/ladder/PositionTest.java
@@ -1,5 +1,6 @@
-package model;
+package model.ladder;
 
+import model.Direction;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
