+++ src/test/java/model/people/PersonTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.people;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
