+++ src/test/java/model/prizes/PrizeTest.java
@@ -1,4 +1,4 @@
-package model;
+package model.prizes;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
