+++ src/test/java/model/prizes/RewardBoardTest.java
@@ -1,9 +1,10 @@
-package model;
+package model.prizes;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import java.util.List;
+import model.people.Person;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
